fix all OpenBTS-UMTS build warnings on gcc-7.2.0

+++ CommonLibs/Utils.cpp
@@ -195,7 +195,7 @@ std::string replaceAll(const std::string input, const std::string search, const 
 
 	while (true) {
 		index = output.find(search, index);
-		if (index == std::string::npos) {
+		if (index == (int)std::string::npos) {
 			break;
 		}
